Add Linting CI
Linter is a program that checks and solves for common code problems
like bad pratices, inconsistency, style issues, high complexity,
duplications, etc.

It's a good idea to have linters as a test case for the continuous
integration system.

In this commit, I've fixed the reported linting issues.

For HTMLLint, Optional Tag, Concerns Seperation, and Capitalization
are disabled due to concerns of unexpected behaviors.

While indentation check is disabled because the file uses 4 spaces
rather than 2 spaces and SpaceConsistencyBear does indentation
check out of the box.