Font depends on OS
I’ve made the CSS use the main system font based on the OS of the user.
If the user is using macOS they’ll see San Francisco, if the user uses
Windows, the font will be Segoe UI, if the user uses Chrome OS or
Android they’ll see Roboto, otherwise it’ll use Helvetica or Arial.
This is inspired by GitHub, they do the same thing, use a font based on
the OS you’re using.